Signup and otp Verification with dashboard v(2)
I see this error with the app, reported by NextJS, please fix it. The er

I see this error with the app, reported by NextJS, please fix it. The er

I see this error with the app, reported by NextJS, please fix it. The er

I see this error with the app, reported by NextJS, please fix it. The er

I see this error with the app, reported by NextJS, please fix it. The er

Make the following changes:

,

Find this element:

  <div class="flex i

first show the message otp verified and make it dissapper in 1 sec and t

make a person stay logged in (on our website) until and unless he/she lo

Make the following changes:

,

Find this element:

  <Input type="tel"

Make the following changes:

,

Find this element:

  <Input type="text"

Make the following changes:

,

Find this element:

  <Button className=

Make the following changes:

,

Find this element:

  <Input id="phone-n

Make the following changes:

,

Find this element:

  <Button className=

do same for login page too

Make the following changes:

,

Find this element:

  <VerifyLoginOtpFor